feat: added employee card tab
