+++ src/components/three-resume/Stars.tsx
