+++ src/components/three-resume/index.ts
@@ -1,0 +1,9 @@
+export * from "./three-canvas.tsx";
+export * from "./About";
+export * from "./EmployeeCard";
+export * from "./OrbitControlsProvider";
+export * from "./Skills";
+export * from "./Stars";
+export * from "./Title";
+export * from "./WordCloud";
+export * from "./Word";
